stricker bracket match reqs; better merging of annotations with newlines
Stricter bracket match requirements means we're less likely to
accidentally grab [report_ended] or similar non-PHI tags and
bracket uses in notes.

The merge_flag tells us if the previous annotation has features
that mean we should merge it with the current annotation. This
happens when there is some ambiguity in the intervening text
which prevents us from being certain about the boundary between
the annotations in the original raw document.