initial support for knowtator to cas and brat formmated output
We made some simplifying assumptions about the annotations related
to mapping Section Header annotations in the output cas file. The
type system (set for cassis) would need to be adapted for other
types.  Eventually, this can be made configurable through command
line options when the need arises.

The mapping file is optional.  If a mapping file is provided, then
the source classes or types are mapped (one-to-many) to the target
output.  The target output class is treated as primary and gets
precedence in the output files.  When the source maps to many
target types, the target types are added as a comma delimited
list.  If no mapping is found, then 'Unknown/Unclassified' is
used.

When no mapping file is provided, then the source type system
is passed through.