normalized brat lines need id number in standard
The normalization lines in brat ann files:

N1  Reference T2  Wikipedia:12345   Barack Obama

That 12345 was missing from the previous version of the converter
script.  We set this to zero because we're not actually syncing back
to the external reference.  Could be a good idea to do later once
we have a standard id.