normalize dates in PHI-related i2b2 corpora to be reasonable "recent"
Some of the i2b2 dates are far in the future (e.g., 2104 or 2083).
Some deid systems have internal logic to see if a number is a
"reasonable" date meaning that it is reasonablly recent. The new
script randomly inserts new dates for all years that is finds
in DATE type annotations with a year between 1950 and 2021
(or 50-21, in the case of two digit years). The replacement is
non-deterministic and not hashed to the same year can be replaced
with different numbers. This could be changed to a squashing function
or a hashed function that is consistent, if needed.