added second leg of round from for SDoH: from OMOP CDM NoteNLP back to brat
round trip brat to OMOP CDM to brat successfully scores all but
multiply instantiated relations.  That is, if a single trigger
has more than one type, then only one of the triggers gets added
to and extracted from the E* entries in brat.

Overall, the F1 drops from 1.00 to 0.995. Tobacco takes the biggest
hit with an F1 of 0.961.

Need to patch this in the next release.